style: 코드 컨벤션 적용

+++ src/test/java/step4/domain/winning/WinningStatusTest.java
@@ -47,7 +47,7 @@ class WinningStatusTest {
 
         // when
         WinningStatus winningStatus = WinningStatus.from(expected, matchBonus);
-        int actual = winningStatus.getCountOfMatch();
+        long actual = winningStatus.getCountOfMatch();
 
         // then
         assertThat(actual).isEqualTo(expected);
